Split root/server.js into separate modules
