+++ root/server/response.js
@@ -1,0 +1,85 @@
+// This file is part of MusicBrainz, the open internet music database.
+// Copyright (C) 2015-2017 MetaBrainz Foundation
+// Licensed under the GPL version 2, or (at your option) any later version:
+// http://www.gnu.org/licenses/gpl-2.0.txt
+
+const _ = require('lodash');
+const path = require('path');
+const Raven = require('raven');
+const React = require('react');
+const ReactDOMServer = require('react-dom/server');
+
+const i18n = require('../static/scripts/common/i18n');
+const getCookie = require('../static/scripts/common/utility/getCookie');
+const DBDefs = require('./DBDefs');
+const gettext = require('./gettext');
+
+function pathFromRoot(fpath) {
+  return path.resolve(__dirname, '../../', fpath);
+}
+
+function badRequest(err) {
+  return Buffer.from(JSON.stringify({
+    body: err.stack,
+    content_type: 'text/plain',
+    status: 400,
+  }));
+}
+
+function getResponse(requestBody, context) {
+  let status = 200;
+  let Page;
+  let response;
+
+  Raven.setContext({
+    environment: DBDefs.GIT_BRANCH,
+    tags: {
+      git_commit: DBDefs.GIT_SHA,
+    },
+  });
+
+  if (context.user) {
+    Raven.mergeContext({user: _.pick(context.user, ['id', 'name'])});
+  }
+
+  global.$c = context;
+
+  // We use a separate gettext handle for each language. Set the current handle
+  // to be used for this request based on the given 'lang' cookie.
+  i18n.setGettextHandle(gettext.getHandle(getCookie('lang')));
+
+  try {
+    Page = require(pathFromRoot(requestBody.component));
+  } catch (err) {
+    if (err.code === 'MODULE_NOT_FOUND') {
+      try {
+        Page = require(pathFromRoot('root/main/404'));
+        status = 404;
+      } catch (err) {
+        Raven.captureException(err);
+        return badRequest(err);
+      }
+    } else {
+      Raven.captureException(err);
+      return badRequest(err);
+    }
+  }
+
+  try {
+    response = ReactDOMServer.renderToStaticMarkup(
+      React.createElement(Page, requestBody.props)
+    );
+  } catch (err) {
+    Raven.captureException(err);
+    return badRequest(err);
+  }
+
+  return Buffer.from(JSON.stringify({
+    body: response,
+    content_type: 'text/html',
+    status,
+  }));
+}
+
+exports.badRequest = badRequest;
+exports.getResponse = getResponse;
